deleted:    .gitignore 	deleted:    LICENSE 	deleted:    Prueba/maian.py 	modified:   Servidor/BsaeDePrueba.db 	renamed:    Servidor/DataBase/BsaeDePrueba.sqbpro -> Servidor/BsaeDePrueba.sqbpro 	deleted:    Servidor/Controller/SemaforoController.py 	deleted:    Servidor/Controller/UsuarioController.py 	deleted:    Servidor/DataBase/BsaeDePrueba.db 	modified:   Servidor/DataBase/index.py 	renamed:    Servidor/DataBase/HistorialModel.py -> Servidor/HistorialModel.py 	renamed:    Servidor/DataBase/PermisosMoel.py -> Servidor/PermisosMoel.py 	renamed:    Servidor/DataBase/Permisos_R_Usuarios.py -> Servidor/Permisos_R_Usuarios.py 	renamed:    Servidor/Controller/RootController.py -> Servidor/RootController.py 	new file:   Servidor/SemaforoController.py 	renamed:    Servidor/DataBase/SemaforosModel.py -> Servidor/SemaforosModel.py 	new file:   Servidor/UsuarioController.py 	renamed:    Servidor/DataBase/UsuarioModel.py -> Servidor/UsuarioModel.py 	renamed:    Servidor/DataBase/conexion.py -> Servidor/conexion.py 	renamed:    Servidor/DataBase/config.py -> Servidor/config.py 	new file:   Servidor/index.py 	modified:   Servidor/main.py
