--- Servidor/Controller/SemaforoController.py
@@ -1,3 +1,0 @@
-class SemaforoController:
-    def Manual():
-
