+++ Servidor/UsuarioController.py
@@ -1,5 +1,5 @@
-from DataBase.index import index
-from DataBase.UsuarioModel import UsuarioModel
+from index import index
+from UsuarioModel import UsuarioModel
 
 ind = index()
 class UsuarioContrller:
