+++ Servidor/UsuarioModel.py
@@ -1,4 +1,4 @@
-from DataBase.conexion import Data
+from conexion import Data
 
 class UsuarioModel:
     def __init__(self):
